Refactor
- Continue refactor.
- Rename GameContainer to GameEngine and separate Game presentational component from it.
