+++ src/navigator.js
@@ -1,13 +1,13 @@
 import {Home} from "./components/home/Home";
 import {StackNavigator} from "react-navigation";
-import GameContainer from "./containers/GameContainer";
 import LevelSelection from "./components/game/LevelSelection";
+import GameEngine from "./containers/GameEngine";
 
 export const Navigator = StackNavigator(
     {
         Home: {screen: Home},
         LevelSelection: {screen: LevelSelection},
-        PlayLevel: {screen: GameContainer},
+        PlayLevel: {screen: GameEngine},
     },
     {
         headerMode: 'screen'
